Conflicts:
	BeebEm5.xcodeproj/project.pbxproj

+++ BeebEm5-Bridging-Header.h
@@ -15,7 +15,7 @@ int beeb_main(int argc, char *argv[]);
 int beeb_end();
 void Exec6502Instruction();
 void beeb_video(int argc, struct PixelData buffer[]);
-void beeb_handlekeys(long eventkind, unsigned short keycode, char charCode);
+void beeb_handlekeys(long eventkind, unsigned long keycode, char charCode);
 
 int beeb_HandleCommand(unsigned int cmdID);
 void beeb_setFilePath(const char* _path);
